feat: add cms config
