+++ build.config/rewrites/development.js
@@ -4,6 +4,10 @@ module.exports = [
     destination: 'https://test-api.tmrwdao.com/api/app/:path*',
   },
   {
+    source: '/cms/:path*',
+    destination: 'https://test.tmrwdao.com/cms/:path*',
+  },
+  {
     source: '/connect/token',
     destination: 'https://test-api.tmrwdao.com/connect/token',
   },
