+++ build.config/rewrites/production.js
@@ -4,6 +4,10 @@ module.exports = [
     destination: 'https://api.tmrwdao.com/api/:path*',
   },
   {
+    source: '/cms/:path*',
+    destination: 'https://tmrwdao.com/cms/:path*',
+  },
+  {
     source: '/connect/token',
     destination: 'https://api.tmrwdao.com/connect/token',
   },
