+++ src/app/telegram/votigram/components/VoteList/index.tsx
@@ -21,6 +21,7 @@ import SignalRManager from 'utils/socket';
 import SignalR from 'utils/socket/signalr';
 import { IPointsListRes, IWsPointsItem } from './type';
 import { preloadImages } from 'utils/file';
+import { useConfig } from 'components/CmsGlobalConfig/type';
 
 interface IVoteListProps {
   onShowMore?: (item: IRankingListResItem) => void;
@@ -51,6 +52,7 @@ export default function VoteList(props: IVoteListProps) {
     setWsRankList(data);
   };
 
+  const { voteMain } = useConfig() ?? {};
   const {
     data: rankList,
     error: rankListError,
@@ -268,7 +270,7 @@ export default function VoteList(props: IVoteListProps) {
         <span className="rule-text">Rules</span>
       </div>
       <h3 className="font-20-25-weight text-white mb-[8px] text-center">
-        🌈 Vote your favorite game!
+        🌈 {voteMain?.listTitle}
       </h3>
       <div className="banner">
         <Carousel autoplay>
@@ -419,16 +421,14 @@ export default function VoteList(props: IVoteListProps) {
         }
       />
       <CommonDrawer
-        title="How to Participate"
+        title={`${voteMain?.rules?.title}`}
         ref={ruleDrawerRef}
         body={
           <div className="flex flex-col items-center">
             <ul className="votigram-rules-text-list">
-              <li>To vote in Votigram, you’ll need a TomorrowPass NFT.</li>
-              <li>You can cast one vote per day and can choose only one option.</li>
-              <li>
-                After voting, you’ll earn points that can be redeemed for exciting rewards later.
-              </li>
+              {voteMain?.rules.description.map((item, index) => {
+                return <li key={index} dangerouslySetInnerHTML={{ __html: item }} />;
+              })}
             </ul>
             <Button
               type="primary"
